refactor(auth): move scope registry to client-safe module, type the dialog's OAuth URL
Review finding on PR 607: getGoogleAuthUrl had zero call sites while
ImportDialog hand-built the identical URL — the duplication that let the
stale type=crm string survive. The helper couldn't simply be imported
into the dialog because googleAuth.ts pulls in ~/server/db at module top.

Split the pure scope registry (GOOGLE_SCOPE_SETS, GOOGLE_SCOPES,
GoogleScopeType, getGoogleAuthUrl) into src/lib/googleScopes.ts with no
server imports; googleAuth.ts re-exports it so server code keeps a single
import path. ImportDialog now calls getGoogleAuthUrl("contacts", ...),
so the scope type is checked by the GoogleScopeType union at compile
time. Redirect URL is byte-identical to the previous template literal.

Bare /api/auth/google-calendar URLs elsewhere (calendar connect buttons)
carry no scope type and rely on the server-side default, so they cannot
drift the same way and are left alone.
